Crate infino

The Infino server application and interface.

The Infino server is an Axum web application that handles all API requests to Infino.

See an Infino architecture overview here. Infino has data ingestion APIs for storing data in Infino and query APIs for retrieving data form Infino. Ingested data is persisted in a queue and forwarded to the CoreDB database that stores and retrieves telemetry data in Infino.

We also summarize logs using Generative AI models; we are currently using OpenAI but we are evaulating alternatives like Llama2 and our own homegrown models. More to come.

Modules

queue_manager 🔒

Manages the ingestion queue for Infino.

utils 🔒

Utilities for managing the Infino application.

Structs

AppState 🔒

Represents application state.

LogsQuery 🔒

Represents a logs query.

MetricsQuery 🔒

Represents a metrics query.

SummarizeQuery 🔒

Represents summarization query. ‘k’ is the number of results to summarize.

SummarizeQueryResponse 🔒

Represents summarization query. ‘k’ is the number of results to summarize.

Functions

app 🔒

Axum application for Infino server.

append_log 🔒

Append log data to CoreDB.

append_metric 🔒

Append metric data to CoreDB.

append_ts 🔒 Deprecated

Deprecated function for backwards-compatibility. Wraps append_metric().

commit_in_loop 🔒

Periodically commits coredb (typically called in a thread, so that coredb can be asyncronously committed).

create_index 🔒

Create a new index in CoreDB with the given name.

delete_index 🔒

Create a function to delete an index with given name.

flush 🔒

Flush the index to disk.

get_index_dir 🔒

Get index directory used by CoreDB.

get_timestamp 🔒

Helper function to get timestamp value from given json object.

main 🔒

Program entry point.

parse_json 🔒

Helper function to parse json input.

ping 🔒

Ping to check if the server is up.

search_log 🔒 Deprecated

Deprecated function for backwards-compatibility. Wraps search_logs().

search_logs 🔒

Search logs in CoreDB.

search_metrics 🔒

Search metrics in CoreDB.

search_ts 🔒 Deprecated

Deprecated function for backwards-compatibility. Wraps search_metrics().

summarize 🔒

Search and summarize logs in CoreDB.